stt_service/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Whisper/Transcription
openai-whisper
ctranslate2>=4.0,<5

# PyTorch (CPU version)
torch==2.2.0
torchaudio==2.2.0

# Core Dependencies
numpy==1.24.3
pydantic==2.4.2
python-multipart==0.0.6

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1

# Utilities
aiofiles==23.2.1
python-dotenv==1.0.0
loguru==0.7.2
psutil==5.9.6
pandas==2.1.3
websockets==12.0
python-magic==0.4.27

# Optional: Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

